requirements.txt
torch==2.2.2
diffusers==0.30.0
transformers==4.43.3
accelerate==0.30.1
safetensors==0.4.3
Pillow==10.3.0
numpy<2
imageio>=2.34.0
imageio-ffmpeg>=0.4.9
